pyproject.toml
[project]
name = "simpleai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "pyinstaller>=6.14.2",
    "pyinstaller-hooks-contrib>=2025.8",
    "python-dotenv==1.1.0",
]
